Make locaiton picker global
