+++ lib/features/add_new_workspace/ui/add_new_worksapce.dart
@@ -10,9 +10,9 @@ import 'package:ibm_flutter_final_project/core/routing/routes.dart';
 import 'package:ibm_flutter_final_project/core/theming/styles.dart';
 import 'package:ibm_flutter_final_project/core/widgets/app_text_button.dart';
 import 'package:ibm_flutter_final_project/core/widgets/image_picker.dart';
+import 'package:ibm_flutter_final_project/core/widgets/location_picker.dart';
 import 'package:ibm_flutter_final_project/features/add_new_workspace/logic/AddNewWorkSpaceCubit/add_new_work_space_cubit.dart';
 import 'package:ibm_flutter_final_project/features/add_new_workspace/logic/AddNewWorkSpaceCubit/add_new_work_space_cubit_state.dart';
- import 'package:ibm_flutter_final_project/features/add_new_workspace/ui/widgets/location_picker.dart';
 import 'package:ibm_flutter_final_project/features/add_new_workspace/ui/widgets/textfield_with_label.dart';
 import 'package:ibm_flutter_final_project/features/authentication/data/repos/signup_repo.dart';
 
@@ -101,31 +101,49 @@ class _AddNewWorkspaceState extends State<AddNewWorkspace> {
                     },
                   ),
                   verticalSpace(10.h),
-                 BlocProvider(
-  create: (_) => AddNewWorkSpaceCubit(),
-  child: BlocBuilder<AddNewWorkSpaceCubit, AddNewWorkSpaceState>(
-    builder: (context, state) {
-          log("image is  in screen ${state.imageFile}");
+                  BlocProvider(
+                    create: (_) => AddNewWorkSpaceCubit(),
+                    child:
+                        BlocBuilder<AddNewWorkSpaceCubit, AddNewWorkSpaceState>(
+                      builder: (context, state) {
+                        log("image is  in screen ${state.imageFile}");
 
-      final cubit = context.read<AddNewWorkSpaceCubit>();
-      return CustomImagePicker(
-        title: "Image",
-        onImagePicked: (image) {
-          // Call the cubit's imageChange method with the picked image
-          cubit.imageChange(image);
-        },
-        selectedImage: state.imageFile, // Use the current image from the cubit's state
-      );
-    },
-  ),
-),
+                        final cubit = context.read<AddNewWorkSpaceCubit>();
+                        return CustomImagePicker(
+                          title: "Image",
+                          onImagePicked: (image) {
+                            // Call the cubit's imageChange method with the picked image
+                            cubit.imageChange(image);
+                          },
+                          selectedImage: state
+                              .imageFile, // Use the current image from the cubit's state
+                        );
+                      },
+                    ),
+                  ),
 
                   // Location Picker Widget
-                  LocationPickerWidget(onLocationPicked: (location) {
-                    // Handle location picked
-                  }),
+                  BlocProvider(
+                    create: (_) => AddNewWorkSpaceCubit(),
+                    child:
+                        BlocBuilder<AddNewWorkSpaceCubit, AddNewWorkSpaceState>(
+                      builder: (context, state) {
+                        final cubit = context.read<AddNewWorkSpaceCubit>();
+
+                        return LocationPickerWidget(
+                          onLocationPicked: (pickedLocation) {
+                            final googleMapUrl =
+                                'https://www.google.com/maps?q=${pickedLocation?.latitude},${pickedLocation?.longitude}';
+                            cubit.locationChange(
+                                googleMapUrl); // Use cubit to save the location
+                            print(googleMapUrl); // Debugging log
+                          },
+                        );
+                      },
+                    ),
+                  ),
+
                   verticalSpace(20.h),
-              
 
                   // Create New Workspace Button
                   Padding(
